added (possibly useless) option to use an iterative base-solver in d3f_util_new.lua. The idea is to allow for a parallel BiCGStab or something similar.
Added apps/d3f/nawak/nawak.lua for simulations on the whole domain.

Added 'organic-box' geometry to the skin application. The idea is to provide
a different lipid meshing for the brick-geometry, which is more similar to the
way in which the lipid is meshed in TKD geometries.

Added several new anisotropic refinement strategies. This area is still
work in progress.